Add support for capping a byte.
The motivation here is that we currently have a situation where
we can get "false uniqueness" where two different values drawn
are actually the same, but because of ignored high bits they end
up looking different. The matrix example in test_composite is an
example where this can cause a problem.

By masking off those high bits in the data stream we fix this
problem by being explicti about what we're ignoring.

We use a cap rather than a mask in the implementation because this
makes it easier to count tree nodes and tell when they are dead.

The addition of draw_bits also allows us to prune the size of the
search space in biased_coin by reducing the number of bits needed
when the weight is i / 2^n for small n.